requirements.txt
scikit-learn
pandas
python-binance
pickle
numpy
matplotlib
xgboost
lightgbm